Add in some more type-dependent definitions and define 64-bit atomic functions on powerpc64. This (a) manages to not break a 32-bit kernel and (b) gets all of the MI stuff in the 64-bit kernel compiling. The elf.h changes are a hack and should be revisited after a closer look at the 64-bit ABI.

+++ sys/powerpc/include/db_machdep.h
@@ -42,7 +42,7 @@
 #define BYTE_MSF        (1)
 
 typedef	vm_offset_t	db_addr_t;	/* address - unsigned */
-typedef	int		db_expr_t;	/* expression - signed */
+typedef	intptr_t	db_expr_t;	/* expression - signed */
 
 #define	PC_REGS(regs)	((db_addr_t)kdb_thrctx->pcb_lr)
